requirements.txt
flask
pillow
gunicorn